refactor: unify source generators shared infrastructure via link-compile (Route A)
Eliminate the Observables.SourceGenerators.Shared.dll and migrate all generators
to link-compile shared source files via a new props-based mechanism.

Changes:
- Create Observables.SourceGenerators.SharedSource.props to link-compile shared
  source files (SourceWriter, ImmutableEquatableArray, IsExternalInit,
  GeneratedSourceHeader, INamedTypeSymbolExtensions, ObservableReturnTypeParser)
- Import shared source props in Observables.SourceGenerators.props (all generators)
- Remove 7 copies of SourceWriter.cs, ImmutableEquatableArray.cs, IsExternalInit.cs
  from SignalR/Mqtt/WebSocket/Grpc/Sse/Nats/RestAPI shproj directories
- Remove ProjectReference to shared project from Events and RestAPI generators
- Remove InterfaceProxy.props (superseded by SharedSource.props)
- Remove shared dll packing from PackRoslynAnalyzer.targets
- Remove ObservablesSharedGeneratorProject property from Observables.Package.props
- Delete central Observables.SourceGenerators.Shared.csproj (source files remain)
- Remove project from Observables.slnx

Benefits:
- No analyzer load-path issues (no shared dll to bundle)
- Single canonical source for shared utilities
- Eliminates 21 duplicate files (7 domains x 3 files)

Closes: (to be filled by issue number)
