+++ Observables.Events/Observables.Events.Reactive.SourceGenerators/Observables.Events.Reactive.SourceGenerators.csproj
@@ -7,7 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Observables.Shared\Observables.SourceGenerators.Shared\Observables.SourceGenerators.Shared.csproj" />
     <PackageReference Include="System.Reactive" PrivateAssets="all" />
   </ItemGroup>
