--- Observables.Grpc/Observables.Grpc.SourceGenerators.Shared/IsExternalInit.cs
@@ -1,6 +1,0 @@
-#if NETSTANDARD2_0 || NETSTANDARD2_1
-namespace System.Runtime.CompilerServices
-{
-    internal static class IsExternalInit { }
-}
-#endif
