+++ Observables.Grpc/Observables.Grpc.SourceGenerators.Shared/Observables.Grpc.SourceGenerators.Shared.projitems
@@ -1,26 +1,23 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>e8f4c3b2-9d6a-4f1e-8c7b-2a3f0d9e6b5c</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Observables.Grpc.Generators</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <AdditionalFiles Include="$(MSBuildThisFileDirectory)AnalyzerReleases.Shipped.md" />
-    <AdditionalFiles Include="$(MSBuildThisFileDirectory)AnalyzerReleases.Unshipped.md" />
-    <Compile Include="$(MSBuildThisFileDirectory)DiagnosticDescriptors.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Emitter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ImmutableEquatableArray.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IncrementalValuesProviderExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IsExternalInit.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\ContextGenerationModel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\GrpcBoundaryKind.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\GrpcInterfaceModel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\GrpcMemberModel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Parser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)SourceWriter.cs" />
-  </ItemGroup>
-</Project>
+<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>e8f4c3b2-9d6a-4f1e-8c7b-2a3f0d9e6b5c</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Observables.Grpc.Generators</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <AdditionalFiles Include="$(MSBuildThisFileDirectory)AnalyzerReleases.Shipped.md" />
+    <AdditionalFiles Include="$(MSBuildThisFileDirectory)AnalyzerReleases.Unshipped.md" />
+    <Compile Include="$(MSBuildThisFileDirectory)DiagnosticDescriptors.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Emitter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)IncrementalValuesProviderExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Models\ContextGenerationModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Models\GrpcBoundaryKind.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Models\GrpcInterfaceModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Models\GrpcMemberModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Parser.cs" />
+  </ItemGroup>
+</Project>
