+++ Observables.RestAPI/Observables.RestAPI.SourceGenerators.Shared/Observables.RestAPI.SourceGenerators.Shared.projitems
@@ -13,9 +13,7 @@
     <AdditionalFiles Include="$(MSBuildThisFileDirectory)AnalyzerReleases.Unshipped.md" />
     <Compile Include="$(MSBuildThisFileDirectory)DiagnosticDescriptors.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Emitter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ImmutableEquatableArray.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IncrementalValuesProviderExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IsExternalInit.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ITypeSymbolExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Models\ContextGenerationModel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Models\InterfaceModel.cs" />
@@ -25,7 +23,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Models\WellKnownTypes.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Parser.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Polyfills\IndexRange.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)SourceWriter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UniqueNameBuilder.cs" />
   </ItemGroup>
 </Project>
