+++ Observables.Shared/Observables.SourceGenerators.Shared/IsExternalInit.cs
