--- Observables.Shared/Observables.SourceGenerators.Shared/Observables.SourceGenerators.Shared.csproj
@@ -1,40 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <RootNamespace>Observables.SourceGenerators.Shared</RootNamespace>
-    <Description>Shared Roslyn infrastructure for Observables source generators.</Description>
-    <EnforceExtendedAnalyzerRules>true</EnforceExtendedAnalyzerRules>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <AdditionalFiles Include="AnalyzerReleases.Shipped.md" />
-    <AdditionalFiles Include="AnalyzerReleases.Unshipped.md" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.CodeAnalysis.Analyzers">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <InternalsVisibleTo Include="Observables.Events.R3.SourceGenerators" />
-    <InternalsVisibleTo Include="Observables.Events.R3.SourceGenerators.Tests" />
-    <InternalsVisibleTo Include="Observables.Events.Reactive.SourceGenerators" />
-    <InternalsVisibleTo Include="Observables.Events.Reactive.SourceGenerators.Tests" />
-    <InternalsVisibleTo Include="Observables.RestAPI.Reactive.SourceGenerators" />
-    <InternalsVisibleTo Include="Observables.RestAPI.R3.SourceGenerators" />
-    <InternalsVisibleTo Include="Observables.RestAPI.GeneratorTests" />
-    <InternalsVisibleTo Include="Observables.SignalR" />
-    <InternalsVisibleTo Include="Observables.SignalR.R3" />
-    <InternalsVisibleTo Include="Observables.WebSocket" />
-    <InternalsVisibleTo Include="Observables.WebSocket.R3" />
-    <InternalsVisibleTo Include="Observables.Mqtt" />
-    <InternalsVisibleTo Include="Observables.Mqtt.R3" />
-    <InternalsVisibleTo Include="Observables.Grpc" />
-    <InternalsVisibleTo Include="Observables.Grpc.R3" />
-  </ItemGroup>
-
-</Project>
