--- Observables.SourceGenerators.InterfaceProxy.props
@@ -1,19 +1,0 @@
-<Project>
-
-  <!--
-    Interface-proxy generators (SignalR, Mqtt, WebSocket) compile ObservableReturnTypeParser
-    from source so ProjectReference OutputItemType="Analyzer" consumers do not need a separate
-    Observables.SourceGenerators.Shared.dll on the Roslyn analyzer load path.
-    Events / RestAPI generators keep a ProjectReference to Observables.SourceGenerators.Shared.
-  -->
-  <PropertyGroup>
-    <NoWarn>$(NoWarn);CS1573</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Observables.Shared\Observables.SourceGenerators.Shared\Diagnostics\ObservableReturnTypeParser.cs">
-      <Link>Shared\Diagnostics\ObservableReturnTypeParser.cs</Link>
-    </Compile>
-  </ItemGroup>
-
-</Project>
